requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Babel==2.14.0
backports.zoneinfo==0.2.1
bleach==6.1.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.9
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-datetime-widget==0.9.3
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-money==3.4.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
fontawesomefree==6.5.1
fonttools==4.57.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
lxml==5.3.2
oscrypto==1.3.0
Pillow==10.1.0
py-moneyed==3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
PyPDF2==3.0.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.6
pytz==2023.3.post1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.3.1
requests==2.32.3
segno==1.6.1
six==1.17.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.17
zipp==3.20.2
zopfli==0.2.3.post1
